+++ android/src/main/java/io/parity/signer/screens/initial/WaitingScreen.kt
@@ -1,4 +1,4 @@
-package io.parity.signer.screens.onboarding
+package io.parity.signer.screens.initial
 
 import android.content.res.Configuration
 import androidx.compose.foundation.Image
