+++ android/src/main/java/io/parity/signer/screens/initial/eachstartchecks/EachStartViewModel.kt
@@ -1,16 +1,12 @@
-package io.parity.signer.screens.onboarding.eachstartchecks
+package io.parity.signer.screens.initial.eachstartchecks
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import io.parity.signer.dependencygraph.ServiceLocator
 import io.parity.signer.domain.Authentication
 import io.parity.signer.domain.NetworkState
 import io.parity.signer.domain.RootUtils
-import io.parity.signer.domain.mapState
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.map
 
 
 class EachStartViewModel : ViewModel() {
